Defines shared API contract types
Moves AI-related types to a shared location to serve as the canonical source of truth for API contracts between the Next.js frontend and the Python backend.
This ensures consistency and avoids duplication of type definitions.
